+++ src/main/java/roomescape/repository/reservationTime/JdbcReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.repository.reservationTime;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
